Number of changes to aid ci testing: * Formalised build process with a Makefile * Fixed issues associated with -Wall -Werror * Updated MQTT_BROKER_IP to point to the test.mosquitto.org test broker IP - need to implement gethostbyname

+++ uMQTT.h
@@ -258,7 +258,7 @@ struct mqtt_packet {
 
   struct raw_pkt raw;
 
-  uint8_t fix_len;
+  size_t fix_len;
   size_t var_len;
   size_t pay_len;
   size_t len;
